+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Added Credential Authorizer in all tx gateways to enable API solution for blockchain apps 
 * Removed all Create TX API
 * Removed API Pricing Credits, now all API will take 1 credit
-* Optimized Pricing & Credits 
+* Restructured Pricing & Credits 
 * Renamed Credits to API calls
 
 ### v11.1.0